requirements.txt
asn1crypto==0.24.0
beautifulsoup4==4.8.0
bs4==0.0.1
certifi==2019.6.16
cffi==1.12.3
chardet==3.0.4
cryptography==2.7
future==0.17.1
idna==2.8
pkg-resources==0.0.0
pycparser==2.19
python-telegram-bot==11.1.0
requests==2.22.0
selenium==3.141.0
six==1.12.0
soupsieve==1.9.2
urllib3==1.25.3